[backend] Extract common dashboard utils
