+++ opencti-platform/opencti-graphql/src/modules/workspace/workspace-domain.ts
@@ -1,12 +1,11 @@
 import * as R from 'ramda';
 import type { FileHandle } from 'fs/promises';
-import { v4 as uuidv4 } from 'uuid';
 import pjson from '../../../package.json';
 import { createEntity, deleteElementById, fullEntitiesOrRelationsConnection, pageEntitiesOrRelationsConnection, updateAttribute } from '../../database/middleware';
 import { fullEntitiesList, pageEntitiesConnection, storeLoadById } from '../../database/middleware-loader';
 import { BUS_TOPICS } from '../../config/conf';
 import { delEditContext, notify, setEditContext } from '../../database/redis';
-import { type BasicStoreEntityWorkspace, ENTITY_TYPE_WORKSPACE, type StoreEntityWorkspace, type WidgetConfiguration } from './workspace-types';
+import { type BasicStoreEntityWorkspace, ENTITY_TYPE_WORKSPACE, type StoreEntityWorkspace } from './workspace-types';
 import { DatabaseError, ForbiddenAccess, FunctionalError } from '../../config/errors';
 import type { AuthContext, AuthUser } from '../../types/user';
 import type {
@@ -28,12 +27,12 @@ import type { BasicConnection, BasicStoreBase, BasicStoreEntity } from '../../ty
 import { buildPagination, isEmptyField, isNotEmptyField, READ_DATA_INDICES_WITHOUT_INTERNAL, READ_INDEX_INTERNAL_OBJECTS } from '../../database/utils';
 import { addFilter } from '../../utils/filtering/filtering-utils';
 import { extractContentFrom } from '../../utils/fileToContent';
-import { isCompatibleVersionWithMinimal } from '../../utils/version';
 import { getEntitiesListFromCache } from '../../database/cache';
 import { ENTITY_TYPE_PUBLIC_DASHBOARD, type PublicDashboardCached } from '../publicDashboard/publicDashboard-types';
 import { convertWidgetsIds } from './workspace-utils';
 import { fromB64, toB64 } from '../../utils/base64';
 import { createInternalObject, editInternalObject } from '../../domain/internalObject';
+import { checkDashboardConfigurationImport, exportDashboardWidget, importDashboardWidgetConfiguration } from '../dashboard/dashboard-utils';
 
 export const PLATFORM_DASHBOARD = 'cf093b57-713f-404b-a210-a1c5c8cb3791';
 
@@ -86,7 +85,7 @@ export const workspaceEditAuthorizedMembers = async (
 };
 
 export const getCurrentUserAccessRight = async (
-  context: AuthContext,
+  _context: AuthContext,
   user: AuthUser,
   workspace: BasicStoreEntityWorkspace,
 ) => {
@@ -287,41 +286,6 @@ export const workspaceEditContext = async (
   );
 };
 
-const MINIMAL_COMPATIBLE_VERSION = '5.12.16';
-const configurationImportTypeValidation = new Map<string, string>();
-configurationImportTypeValidation.set(
-  'dashboard',
-  'Invalid type. Please import OpenCTI dashboard-type only',
-);
-configurationImportTypeValidation.set(
-  'widget',
-  'Invalid type. Please import OpenCTI widget-type only',
-);
-configurationImportTypeValidation.set(
-  'theme',
-  'Invalid type. Please import OpenCTI theme-type only',
-);
-
-interface ConfigImportData {
-  type: string;
-  openCTI_version: string;
-}
-
-export const checkConfigurationImport = (type: string, parsedData: ConfigImportData) => {
-  if (configurationImportTypeValidation.has(type) && parsedData.type !== type) {
-    throw FunctionalError(configurationImportTypeValidation.get(type), {
-      reason: parsedData.type,
-    });
-  }
-
-  if (!isCompatibleVersionWithMinimal(parsedData.openCTI_version, MINIMAL_COMPATIBLE_VERSION)) {
-    throw FunctionalError(
-      `Invalid version of the platform. Please upgrade your OpenCTI. Minimal version required: ${MINIMAL_COMPATIBLE_VERSION}`,
-      { reason: parsedData.openCTI_version },
-    );
-  }
-};
-
 // region workspace ids converter_2_1
 // Export => Dashboard filter ids must be converted to standard id
 // Import => Dashboards filter ids must be converted back to internal id
@@ -358,24 +322,16 @@ export const generateWidgetExportConfiguration = async (context: AuthContext, us
   if (workspace.type !== 'dashboard') {
     throw FunctionalError('WORKSPACE_EXPORT_INCOMPATIBLE_TYPE', { type: workspace.type });
   }
-  const parsedManifest = fromB64(workspace.manifest ?? '{}');
-  if (parsedManifest && isNotEmptyField(parsedManifest.widgets) && parsedManifest.widgets[widgetId]) {
-    const widgetDefinition = parsedManifest.widgets[widgetId];
-    delete widgetDefinition.id; // Remove current widget id
-    await convertWidgetsIds(context, user, [widgetDefinition], 'internal');
-    const exportConfigration = {
-      openCTI_version: pjson.version,
-      type: 'widget',
-      configuration: toB64(widgetDefinition) as string,
-    };
-    return JSON.stringify(exportConfigration);
+  const result = await exportDashboardWidget(context, user, workspace.manifest, widgetId);
+  if (!result.success) {
+    throw FunctionalError('WIDGET_EXPORT_NOT_FOUND', { workspace: workspace.id, widget: widgetId });
   }
-  throw FunctionalError('WIDGET_EXPORT_NOT_FOUND', { workspace: workspace.id, widget: widgetId });
+  return result.data;
 };
 
 export const workspaceImportConfiguration = async (context: AuthContext, user: AuthUser, file: Promise<FileHandle>) => {
   const parsedData = await extractContentFrom(file);
-  checkConfigurationImport('dashboard', parsedData);
+  checkDashboardConfigurationImport('dashboard', parsedData);
   const authorizedMembers = initializeAuthorizedMembers([], user);
   const { manifest } = parsedData.configuration;
   // Manifest ids must be rewritten for filters
@@ -421,49 +377,18 @@ export const duplicateWorkspace = async (context: AuthContext, user: AuthUser, i
   return notify(BUS_TOPICS[ENTITY_TYPE_WORKSPACE].ADDED_TOPIC, created, user);
 };
 
-interface WidgetConfigImportData extends ConfigImportData {
-  configuration?: string; // widget definition in base64.
-}
-
 export const workspaceImportWidgetConfiguration = async (
   context: AuthContext,
   user: AuthUser,
   workspaceId: string,
   input: ImportWidgetInput,
 ) => {
-  const parsedData = await extractContentFrom<WidgetConfigImportData>(input.file);
-  checkConfigurationImport('widget', parsedData);
-  const widgetDefinition = fromB64(parsedData.configuration);
-  await convertWidgetsIds(context, user, [widgetDefinition], 'stix');
-  const importedWidgetId = uuidv4();
-  const dashboardManifestObjects = fromB64(input.dashboardManifest ?? undefined);
-
-  // When importing a widget, change its position to not break
-  // the current layout of the dashboard.
-  // It is moved on a new line.
-  const widgetsArray = Object.values(dashboardManifestObjects.widgets ?? [])
-    .map((widget) => widget) as WidgetConfiguration[];
-  const nextRow = widgetsArray.reduce((max, { layout }) => {
-    const widgetEndRow = layout.y + layout.h;
-    return widgetEndRow > max ? widgetEndRow : max;
-  }, 0);
-
-  const updatedObjects = {
-    ...dashboardManifestObjects,
-    widgets: {
-      ...dashboardManifestObjects.widgets,
-      [importedWidgetId]: {
-        id: importedWidgetId,
-        ...widgetDefinition,
-        layout: {
-          ...widgetDefinition.layout,
-          x: 0,
-          y: nextRow,
-        },
-      },
-    },
-  };
-  const updatedManifest = toB64(updatedObjects);
+  const { updatedManifest, importedWidgetId } = await importDashboardWidgetConfiguration(
+    context,
+    user,
+    input.file,
+    input.dashboardManifest,
+  );
   const { element } = await updateAttribute<StoreEntityWorkspace>(
     context,
     user,
